Moved the modules to a package magic/
Changed the import statements in the modules inside the package for
intra-package-references
See https://docs.python.org/3/tutorial/modules.html#intra-package-references

Also imported the makeExtension function of the magic module to
__init__.py so that extension users can still just use 'magic' as the
extension name (and not magic.magic or magic.magic:Magic)
