+++ magic/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ['magic', 'imageGen', 'gifAPI', 'memeAPI']
+
+from .magic import makeExtension
