+++ setup.py
@@ -7,8 +7,8 @@ setup(
     maintainer='Radhika Ghosal',
     maintainer_email='radhikaghosal@gmail.com',
     url='https://github.com/RadhikaG/markdown-magic',
-    py_modules=['magic'],
-    install_requires=['markdown>=2.5'],
+    packages=['magic'],
+    install_requires=['markdown>=2.5', 'requests'],
     license='MIT',
     description='Create memes and find best-fitting gifs using Python-Markdown.',
     keywords=['markdown','markup','memes','gifs','plugin'],
